+++ src/main/java/in/zapr/druid/druidry/filter/havingSpec/NotHaving.java
@@ -1,4 +1,4 @@
-package in.zapr.druid.druidry.having;
+package in.zapr.druid.druidry.filter.havingSpec;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
